refactor: renames, add browser test debuggers
