+++ src/lib/simulations/tests/dpsScore/dpsScoreOrchestratorTestUtils.ts
@@ -1,5 +1,5 @@
 import { runDpsScoreBenchmarkOrchestrator } from 'lib/simulations/orchestrator/runDpsScoreBenchmarkOrchestrator'
-import { generateTestSingleRelicsByPart, TestInput } from 'lib/simulations/tests/statSim/statSimTestUtils'
+import { generateTestSingleRelicsByPart, TestInput } from 'lib/simulations/tests/simTestUtils'
 import DB from 'lib/state/db'
 import { TsUtils } from 'lib/utils/TsUtils'
 import { Character } from 'types/character'
